fix: just use project as scope
